+++ ics_tender_management/models/__init__.py
@@ -1,6 +1,7 @@
 from . import tender_stage
 from . import tender
 from . import tender_boq
+from . import tender_dashboard
 from . import crm_lead
 from . import purchase_requisition
 from . import sale_order